css/css-position/position-sticky-right.html

<!DOCTYPE html>
<title>position:sticky elements should respect the right constraint</title>
<link rel="help" href="https://www.w3.org/TR/css-position-3/#sticky-pos" />
<meta name="assert" content="This test checks that position:sticky elements obey their right anchor after scrolling" />

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
.scroller {
  position: relative;
  width: 200px;
  height: 100px;
  overflow-x: auto;
  overflow-y: hidden;
}

.contents {
  height: 100%;
  width: 500px;
}

.prepadding {
  display: inline-block;
  height: 100%;
  width: 200px;
}

.container {
  display: inline-block;
  height: 100%;
  width: 200px;
}

.filler {
  display: inline-block;
  height: 100%;
  width: 100px;
}

.indicator {
  background-color: red;
  position: absolute;
  top: 0;
}

.sticky {
  background-color: green;
  position: sticky;
  right: 25px;
}

.box {
  display: inline-block;
  height: 100%;
  width: 100px;
}
</style>

<div id="scroller1" class="scroller">
  <div class="indicator box" style="left: 200px;"></div>
  <div class="contents">
    <!-- As these elements are inline, they are whitespace sensitive. -->
    <div class="prepadding"></div><div class="container"><div class="filler"></div><div class="sticky box"></div></div>
  </div>
</div>

<div id="scroller2" class="scroller">
  <div class="indicator box" style="left: 250px;"></div>
  <div class="contents">
    <!-- As these elements are inline, they are whitespace sensitive. -->
    <div class="prepadding"></div><div class="container"><div class="filler"></div><div class="sticky box"></div></div>
  </div>
</div>

<div id="scroller3" class="scroller">
  <div class="indicator box" style="left: 300px;"></div>
  <div class="contents">
    <!-- As these elements are inline, they are whitespace sensitive. -->
    <div class="prepadding"></div><div class="container"><div class="filler"></div><div class="sticky box"></div></div>
  </div>
</div>

<script>
test(() => {
  const scroller = document.getElementById('scroller1');
  const prepadding = scroller.querySelector('.prepadding');
  const filler = scroller.querySelector('.filler');
  const sticky = scroller.querySelector('.sticky');

  scroller.scrollLeft = 250;

  const nonStickyLeftY = prepadding.clientWidth + filler.clientWidth;
  assert_equals(sticky.offsetLeft, nonStickyLeftY);
}, 'before reaching the sticking point the sticky box should not be offset');

test(() => {
  const scroller = document.getElementById('scroller2');
  const prepadding = scroller.querySelector('.prepadding');
  const filler = scroller.querySelector('.filler');
  const sticky = scroller.querySelector('.sticky');

  scroller.scrollLeft = 175;

  const nonStickyLeftY = prepadding.clientWidth + filler.clientWidth;
  const nonStickyBottomY = nonStickyLeftY + sticky.clientWidth;
  const targetBottomY = scroller.clientWidth + scroller.scrollLeft - 25;
  const stickyOffset = nonStickyBottomY - targetBottomY;

  assert_equals(sticky.offsetLeft, nonStickyLeftY - stickyOffset);
}, 'after reaching the sticking point the sticky box should be offset');

test(() => {
  const scroller = document.getElementById('scroller3');
  const container = scroller.querySelector('.container');
  const prepadding = scroller.querySelector('.prepadding');
  const filler = scroller.querySelector('.filler');
  const sticky = container.querySelector('.sticky');

  scroller.scrollLeft = 100;

  // The sticky is at the right, so hax.
  const nonStickyLeftY = prepadding.clientWidth + filler.clientWidth;
  const maxOffsetInContainer = container.clientWidth - sticky.clientWidth;

  assert_equals(sticky.offsetLeft, nonStickyLeftY - maxOffsetInContainer);
}, 'the sticky box should not be pushed outside its containing block');
</script>
